+++ src/Commands/EjectController.php
@@ -23,7 +23,7 @@ use function sprintf;
  * @author Wilmer Arambula <terabytesoftw@gmail.com>
  * @since 0.1
  */
-final class EjectController extends Controller
+class EjectController extends Controller
 {
     /**
      * When `true`, performs the ejection without prompting for confirmation.